+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support for IBM Fortran compiler by circumventing ICE
+
 ## [0.9.3] - 2020-08-02
 
 ### Added